+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * slot for fragment
 * exported fragments (inverted slots)
 * <malina:head>, <malina:body>, <malina:window>
+* option: passClass
 
 ## 0.5.x